Restructure the transformer
